Conflicts:
	src/tilda/db/stores/PostgreSQL.java

+++ src/tilda/types/Type_BooleanCollection.java
@@ -20,7 +20,7 @@ import tilda.enums.ColumnType;
 
 public class Type_BooleanCollection extends ColumnDefinition
   {
-    public Type_BooleanCollection(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_BooleanCollection(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, ColumnType.BOOLEAN, true, Description);
       }
