+++ src/tilda/types/Type_DatetimePrimitiveNull.java
@@ -19,7 +19,7 @@ package tilda.types;
 
 public class Type_DatetimePrimitiveNull extends Type_DatetimePrimitive implements Nullable
   {
-    public Type_DatetimePrimitiveNull(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_DatetimePrimitiveNull(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, Description);
       }
