+++ src/tilda/types/Type_StringCollection.java
@@ -20,7 +20,7 @@ import tilda.enums.ColumnType;
 
 public class Type_StringCollection extends ColumnDefinition
   {
-    public Type_StringCollection(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_StringCollection(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, ColumnType.STRING, true, Description);
       }
