+++ src/tilda/types/Type_StringPrimitiveNull.java
@@ -19,7 +19,7 @@ package tilda.types;
 
 public class Type_StringPrimitiveNull extends Type_StringPrimitive implements Nullable
   {
-    public Type_StringPrimitiveNull(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_StringPrimitiveNull(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, Description);
       }
